backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
Mastodon.py>=1.8.1
python-dotenv>=1.0.0
python-dateutil>=2.8.0